+++ sources/packages/web/src/services/http/dto/SupportingUser.dto.ts.ts
@@ -1,18 +1,13 @@
-import { IsNotEmpty, IsNotEmptyObject, IsOptional } from "class-validator";
-import { SupportingUserType } from "../../../database/entities";
-import { ContactInfo } from "../../../types";
+import { ContactInformation, SupportingUserType } from "@/types";
 
 /**
  * Information used to uniquely identify a Student Application.
  * The application must be search using at least 3 criteria as
  * per defined by the Ministry policies.
  */
-export class ApplicationIdentifierApiInDTO {
-  @IsNotEmpty()
+export interface ApplicationIdentifierApiInDTO {
   applicationNumber: string;
-  @IsNotEmpty()
-  studentsDateOfBirth: string;
-  @IsNotEmpty()
+  studentsDateOfBirth: Date;
   studentsLastName: string;
 }
 
@@ -21,46 +16,33 @@ export class ApplicationIdentifierApiInDTO {
  * The validation of the entire model will (and
  * must) be done by the Form.IO dry run.
  */
-export class UpdateSupportingUserApiInDTO {
-  @IsNotEmpty()
-  applicationNumber: string;
-  @IsNotEmpty()
+export interface UpdateSupportingUserApiInDTO
+  extends ApplicationIdentifierApiInDTO {
   addressLine1: string;
-  @IsOptional()
   addressLine2: string;
-  @IsNotEmpty()
   city: string;
-  @IsNotEmpty()
   country: string;
-  @IsNotEmpty()
   phone: string;
-  @IsNotEmpty()
   postalCode: string;
-  @IsNotEmpty()
   provinceState: string;
-  @IsNotEmpty()
   sin: string;
-  @IsNotEmpty()
-  studentsDateOfBirth: string;
-  @IsNotEmpty()
-  studentsLastName: string;
-  @IsNotEmptyObject()
   supportingData: any;
 }
 
-export class ApplicationApiOutDTO {
+export interface ApplicationApiOutDTO {
   programYearStartDate: Date;
   formName: string;
 }
 
-export class ApplicationSupportingUsersApiOutDTO {
+export interface ApplicationSupportingUsersApiOutDTO {
   supportingUserId: number;
   supportingUserType: SupportingUserType;
 }
-export class SupportingUserFormDataApiOutDTO {
+
+export interface SupportingUserFormDataApiOutDTO {
   formName: string;
   supportingData: any;
-  contactInfo: ContactInfo;
+  contactInfo: ContactInformation;
   sin: string;
   birthDate: Date;
   gender: string;
